Fixed price input to accept float values

+++ database/migrations/2023_02_22_163142_create_products_table.php
@@ -21,8 +21,8 @@ return new class extends Migration
             $table->mediumText('small_description')->nullable();
             $table->longText('description')->nullable();
 
-            $table->integer('original_price');
-            $table->integer('selling_price');
+            $table->float('original_price');
+            $table->float('selling_price');
             $table->integer('quantity');
             $table->tinyInteger('trending')->default('0')->comment('1=trending,0=normal');
             $table->tinyInteger('status')->default('0')->comment('1=hidden,0=visible');
